Adds a bit of documentation to answerChecker
I was recording my stuff on the google doc, and I needed to check here for the function names I had used, and then I realized it could use some more documentation. Also changed the response to an equivalent answer that shouldn't be simplified, as it would say 'simplification is needed' even if it needed to be expanded/factored/etc. This is a long summary.