Finished Website
All four pages are complete for the midterm check.  The CSS was updated as well.  Currently, the webpage works in Chromium and Mozilla-based browsers, as well as Microsoft Edge.  Opera compatibility is questionable and it is not compatible with IE.  If you have any updates to make to the webpage, please let me know.